gitignore file updated and some classes moved to new package "human" .
